% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycont.R
\name{bycont}
\alias{bycont}
\title{Table of one or more categorical variables by a single continuous variable}
\usage{
bycont(catvars, contvar, dat, pval = TRUE)
}
\arguments{
\item{catvars}{is a list of the categorical variables for the rows of the
table e.g. list('Gene1', 'Gene2')}

\item{contvar}{is the continuous variable you want summarized by each
categorical variable. Must be in quotes.}

\item{dat}{is the dataset to use for analysis}

\item{pval}{takes the value TRUE or FALSE indicating whether p-values
should be computed. Defaults to TRUE. When TRUE, Kruskal-Wallis p-values
are produced.}
}
\value{
Returns a dataframe
}
\description{
\code{bycont} takes a list of categorical variabls and returns
median(min, max) of the single continuous variable within each level
of each categorical variable. Computes Kruskall-Wallis p-values.
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
