% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_color_palette.R
\docType{data}
\name{ccf_colors}
\alias{ccf_colors}
\title{Named list of CCF brand colors}
\format{
An object of class \code{character} of length 24.
}
\usage{
ccf_colors
}
\description{
This is named vector of the CCF brand color hex codes. Run \code{ccf_colors} to
see the names and hex codes for the CCF brand colors.
}
\keyword{datasets}
