% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_color_palette.R
\name{ccf_cols}
\alias{ccf_cols}
\title{Function to extract colors from \code{ccf_colors} as hex codes}
\usage{
ccf_cols(...)
}
\arguments{
\item{...}{Character names of ccf_colors}
}
\description{
Function to extract colors from \code{ccf_colors} as hex codes
}
