% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_color_palette.R
\name{ccf_pal}
\alias{ccf_pal}
\title{Return function to interpolate a CCF color palette}
\usage{
ccf_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ccf_palettes.
Options include "main" (default), "bright", "neutral", "all", "blues" and
"greens".}

\item{reverse}{Boolean indicating whether the palette should be reversed.
Default to FALSE.}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a CCF color palette
}
