% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_color_palette.R
\docType{data}
\name{ccf_palettes}
\alias{ccf_palettes}
\title{CCF branded color palettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
ccf_palettes
}
\description{
Creates color palettes based on CCF brand colors, including
"main", which contains all primary and secondary CCF colors,
"bright", which contains all supporting bright CCF colors,
"neutral", which contains all supporting neutral CCF colors,
"all", which contains all colors from main, bright, and neutral for when
more colors are needed,
"blues", which contains a high and low blue color to create continuous blue
scales,
"greens", which contains a high and low green color to create continuous
green scales
}
\examples{
ccf_palettes[["main"]]
ccf_palettes[["contrast"]]
}
\keyword{datasets}
