% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_age.R
\name{compute_age}
\alias{compute_age}
\title{Calculate age in years based on two dates and their respective formats}
\usage{
compute_age(start_date, end_date, data, start_format = "\%m/\%d/\%Y",
  end_format = "\%m/\%d/\%Y")
}
\arguments{
\item{start_date}{is the name of the starting date variable,
supplied in quotes, e.g. "birth_date"}

\item{end_date}{is the name of the ending date variable,
supplied in quotes, e.g. "surgery_date"}

\item{data}{is the dataframe in which to find the start and
end dates}

\item{start_format}{is the format of the character variable
stored in \code{start_date}. The variable will be temporarily
formatted as a date to perform the calculation using
\code{as.Date()}. Defaults to "%m/%d/%Y".
Must be supplied in quotes.
#' @param end_format is the format of the character variable
stored in \code{end_date}. The variable will be temporarily
formatted as a date to perform the calculation using
\code{as.Date()}. Defaults to "%m/%d/%Y".
Must be supplied in quotes}
}
\value{
Returns a continuous integer valued age in years
}
\description{
\code{compute_age} takes two dates, and optionally two date
formats as arguments, and returns a whole number age
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
