% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_template-data.R
\docType{data}
\name{ez_analysis_template}
\alias{ez_analysis_template}
\title{EZ analysis project template}
\format{
A list
}
\usage{
ez_analysis_template
}
\description{
EZ analysis project template
}
\keyword{datasets}
