% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezcolors.R
\name{ez_pal}
\alias{ez_pal}
\title{A custom color scale made by Emily Zabor, with help from Coolors app}
\usage{
ez_pal()
}
\description{
Basically copying code from hrbrmstr/hrbrthemes colors.r
}
\examples{
library(scales)
scales::show_col(ez_pal()(9))
}
