% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowerchar.R
\name{lowerchar}
\alias{lowerchar}
\title{Convert to lowercase}
\usage{
lowerchar(dfname)
}
\arguments{
\item{dfname}{is the name of the dataframe on which to perform the action}
}
\value{
Nothing is returned from \code{lowerchar}, the action is simply perfomed on the columns of
dataframe \code{dfname}
}
\description{
\code{lowerchar} converts the levels of character variables from upper or mixed
case to lower case
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
