% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_color_palette.R
\docType{data}
\name{msk_colors}
\alias{msk_colors}
\title{Named list of MSK brand colors}
\format{An object of class \code{character} of length 20.}
\usage{
msk_colors
}
\description{
This is named vector of the MSK brand color hex codes. Run `msk_colors` to
see the names and hex codes for the MSK brand colors.
}
\keyword{datasets}
