% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_color_palette.R
\name{msk_cols}
\alias{msk_cols}
\title{Function to extract MSK colors as hex codes}
\usage{
msk_cols(...)
}
\arguments{
\item{...}{Character names of msk_colors}
}
\description{
Function to extract MSK colors as hex codes
}
