% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_color_palette.R
\name{msk_pal}
\alias{msk_pal}
\title{Return function to interpolate a MSK color palette}
\usage{
msk_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in msk_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a MSK color palette
}
