% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_color_palette.R
\docType{data}
\name{msk_palettes}
\alias{msk_palettes}
\title{Complete list of MSK branded color palettes}
\format{An object of class \code{list} of length 5.}
\usage{
msk_palettes
}
\description{
Complete list of MSK branded color palettes
}
\keyword{datasets}
