% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvcoxres.R
\name{mvcoxres}
\alias{mvcoxres}
\title{Format resuls from multivariable Cox regression model}
\usage{
mvcoxres(mod)
}
\arguments{
\item{mod}{is a multivariable Cox regression object}
}
\value{
Returns a dataframe
}
\description{
\code{mvcoxres} takes a multivariable Cox regression object and formats the
resulting HR (95\% CI) and p-values into a table
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
