% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvcrrres.R
\name{mvcrrres}
\alias{mvcrrres}
\title{Format resuls from multivariable competing risks regression model}
\usage{
mvcrrres(mod)
}
\arguments{
\item{mod}{is a multivariable Cox regression object}
}
\value{
Returns a dataframe
}
\description{
\code{mvcrrres} takes a multivariable competing risks regression object and puts the
resulting HR (95\% CI) and p-values into a table
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}

