% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlmres.R
\name{mvlmres}
\alias{mvlmres}
\title{Format resuls from multivariable linear regression model}
\usage{
mvlmres(mod)
}
\arguments{
\item{mod}{is a multivariable linear regression object from \code{lm}}
}
\value{
Returns a dataframe
}
\description{
\code{mvlmres} takes a multivariable linear regression object and formats the
resulting Est (SE) and p-values into a table
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
