% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlogitres.R
\name{mvlogitres}
\alias{mvlogitres}
\title{Format resuls from multivariable logistic regression model}
\usage{
mvlogitres(mod)
}
\arguments{
\item{mod}{is a multivariable logistic regression object from \code{glm}}
}
\value{
Returns a dataframe
}
\description{
\code{mvlogitres} takes a multivariable logistic regression object and formats the
resulting OR (95\% CI) and p-values into a table
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
