% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2simonApp.R
\name{ph2simonApp}
\alias{ph2simonApp}
\title{Interactive Simon's 2-stage Shiny app}
\usage{
ph2simonApp()
}
\value{
The output includes 1) fields to enter the design parameters for the
Simon 2-stage Phase II design, 2) R output with \code{\link[clinfun]{ph2simon}} results,
3) a paragraph interpreting the results, and 4) a plot of maximum versus expected
number of patients indicating the optimal and minimax results.
}
\description{
\code{ph2simonApp} is simply a Shiny interface for the
\code{\link[clinfun]{ph2simon}} function. No arguments need to be passed to
the function.
}
\examples{
ph2simonApp()
}
