% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezcolors.R
\name{scale_colour_ez}
\alias{scale_colour_ez}
\alias{scale_color_ez}
\alias{scale_fill_ez}
\title{Discrete color & fill scales based on the ez palette}
\usage{
scale_colour_ez(...)

scale_color_ez(...)

scale_fill_ez(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
See \url{ez_pal}.
}
