% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1_re.R
\name{tab1_re}
\alias{tab1_re}
\title{Table 1 with random effects model p-values}
\usage{
tab1_re(contvars, catvars, byvar, re, dat, col = TRUE,
  spread = "range")
}
\arguments{
\item{contvars}{is a list of the continuous variables you want in the rows
e.g. list('Age'). Can be NULL.}

\item{catvars}{is a list of the categorical variables you want in the rows
e.g. list('Gender','Race'). Can be NULL.}

\item{byvar}{is the categorical variable you want to tabulate by across the
columns (needs to be in quotes). MUST BE 0/1 since it will be used as the
outcome variable in \code{\link[lme4]{glmer}}.}

\item{re}{the name of the random effect variable, should be supplied in quotes, e.g. re = "ptid"}

\item{dat}{is the dataset to use for analysis}

\item{col}{takes the value TRUE or FALSE indicating whether you want column
percent (TRUE, default) or row percent (FALSE)}

\item{spread}{takes the value "range" or "iqr" indicating whether you want
(min, max) or (Q1, Q3) in summaries of continuous variables. Defaults to
"range".}
}
\value{
Returns a dataframe. If there are warnings or errors from
\code{\link[lme4]{glmer}} then NA is returned in place of the p-value.
}
\description{
\code{tab1_re} takes lists of continuous and/or categorical variables and
returns Median (spread) for continuous variables and N (\%) for categorical
variables. Produces a table with both an overall column and columns by
another variable. For a binary by variable only, it produces p-values from a
random effects model.
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
