% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabna.R
\name{tabna}
\alias{tabna}
\title{Cross-tabulation with \code{useNA = "ifany"}}
\usage{
tabna(...)
}
\arguments{
\item{...}{the function takes any standard arguments to \code{table}}
}
\value{
\code{tabna} returns a contingency table with NAs included, if any
}
\description{
\code{tabna} is an implementation of \code{table} with argument \code{useNA = "ifany"}
}
\details{
See the help file for \code{table} for detailed information about possible arguments to the function
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
