% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todate.R
\name{todate}
\alias{todate}
\title{Convert to date format}
\usage{
todate(dfname)
}
\arguments{
\item{dfname}{is the name of the dataframe on which to perform the action}
}
\value{
Nothing is returned from \code{todate}, the action is simply perfomed on the columns of
dataframe \code{dfname}
}
\description{
\code{todate} converts any POSIXct format variables in the dataframe to date format
}
\details{
Note that this function will mainly apply to dataframes imported using the \code{read_excel} function
from the \code{readxl} package. Dataframes imported using, for example, \code{read.csv} instead
will have dates in character format and therefore \code{todate} will not apply.
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
