% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvcoxph.R
\name{uvcoxph}
\alias{uvcoxph}
\title{Table of univariable Cox regression results}
\usage{
uvcoxph(contvars, catvars, event, time, dat, strata = NULL)
}
\arguments{
\item{contvars}{is a list of the continuous variables you want in the rows e.g. list('Age')}

\item{catvars}{is a list of the categorical variables you want in the rows e.g. list('Gender','Race')}

\item{event}{is the event indicator (needs to be in quotes)}

\item{time}{is the survival time variables (needs to be in quotes)}

\item{dat}{is the dataset for analysis}

\item{strata}{is a possible strata term for use in calculating the log-rank
p-values. Defaults to NULL. Entries should be in quotes, e.g. "Surgeon"}
}
\value{
Returns a dataframe. If there are warnings or errors from \code{coxph} then blank rows are returned.
}
\description{
\code{uvcoxph} takes lists of continuous and/or categorical variables, runs a univariable \code{coxph} model for
each, and puts the resulting HR (95\% CI) and p-value into a table suitable for printing in a Word \code{R Markdown}
file.
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
