% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvcrr.R
\name{uvcrr}
\alias{uvcrr}
\title{Table of univariable competing risks regression results}
\usage{
uvcrr(contvars, catvars, event, time, dat)
}
\arguments{
\item{contvars}{is a list of the continuous variables you want in the rows
e.g. list('Age')}

\item{catvars}{is a list of the categorical variables you want in the rows
e.g. list('Gender','Race')}

\item{event}{is the event indicator (needs to be in quotes)}

\item{time}{is the survival time variables (needs to be in quotes)}

\item{dat}{is the dataset for analysis}
}
\value{
Returns a dataframe. If there are warnings or errors from \code{crr}
then blank rows are returned.
}
\description{
\code{uvcrr} takes lists of continuous and/or categorical variables, runs a
univariable \code{crr} model for each, and puts the resulting HR (95\% CI)
and p-value into a table suitable for printing in a Word \code{R Markdown}
file.
}
\details{
\code{uvcrr} uses all function defaults to \code{crr}. For example, the
failure code is set to 1. See the help file for \code{crr} for additional
details.
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
