% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvcrrSC.R
\name{uvcrrSC}
\alias{uvcrrSC}
\title{Table of univariable competing risks regression results}
\usage{
uvcrrSC(contvars, catvars, cluster, event, time, dat)
}
\arguments{
\item{contvars}{is a list of the continuous variables you want in the rows
e.g. list('Age')}

\item{catvars}{is a list of the categorical variables you want in the rows
e.g. list('Gender','Race')}

\item{cluster}{is the variable in the dataset used for clustering (needs to be in quotes e.g. 'mrn')}

\item{event}{is the event indicator (needs to be in quotes)}

\item{time}{is the survival time variables (needs to be in quotes)}

\item{dat}{is the dataset for analysis}
}
\value{
Returns a dataframe. If there are warnings or errors from \code{\link[cmprsk]{crr}}
then blank rows are returned.
}
\description{
\code{uvcrrSC} takes lists of continuous and/or categorical variables, runs a
univariable \code{\link[crrSC]{crrc}} model for each, and puts the resulting HR (95\% CI)
and p-value into a table suitable for printing in a Word R Markdown
file.
}
\details{
\code{uvcrr} uses all function defaults to \code{\link[crrSC]{crrc}}. For example, the
failure code is set to 1. See the help file for \code{\link[crrSC]{crrc}} for additional
details.
}
\author{
Emily Zabor/Margie Hannum \email{hannumm@mskcc.org}
}
