% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvlm.R
\name{uvlm}
\alias{uvlm}
\title{Table of univariable linear regression results}
\usage{
uvlm(contvars, catvars, out, dat)
}
\arguments{
\item{contvars}{is a list of the continuous variables you want in the rows
e.g. list('Age')}

\item{catvars}{is a list of the categorical variables you want in the rows
e.g. list('Gender','Race')}

\item{out}{is the continuous outcome variable (needs to be in quotes)}

\item{dat}{is the dataset for analysis}
}
\value{
Returns a dataframe. If there are warnings or errors from \code{lm}
then blank rows are returned.
}
\description{
\code{uvlm} takes lists of continuous and/or categorical variables, calls
\code{lm} to run a linear regression model for each, and returns a table with
Est (SE) and p-value for each variable that is suitable for printing in a
Word \code{R Markdown} file.
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}

