% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvlogit.R
\name{uvlogit}
\alias{uvlogit}
\title{Table of univariable logistic regression results}
\usage{
uvlogit(contvars, catvars, out, dat)
}
\arguments{
\item{contvars}{is a list of the continuous variables you want in the rows
e.g. list('Age')}

\item{catvars}{is a list of the categorical variables you want in the rows
e.g. list('Gender','Race')}

\item{out}{is the binary outcome variable. Must be coded 0/1.
(needs to be in quotes)}

\item{dat}{is the dataset for analysis}
}
\value{
Returns a dataframe. If there are warnings or errors from \code{\link[stats]{glm}}
then blank rows are returned.
}
\description{
\code{uvlogit} takes lists of continuous and/or categorical variables, calls
\code{\link[stats]{glm}} to run a logistic regression model for each, and returns a table with
OR (95% CI) and p-value for each variable that is suitable for printing in a
Word \code{R Markdown} file.
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
