% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_color_palette.R
\name{msk_palette}
\alias{msk_palette}
\title{MSK branded colors palette generator}
\usage{
msk_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are: \code{all},
\code{main} (default),  \code{blues}, \code{oranges}, \code{contrast}}

\item{n}{Number of colors desired. If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
  @importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
MSK branded colors palette generator
}
\examples{
msk_palette("main")
msk_palette("blues")
msk_palette("contrast")

library(ggplot2)

# use a single brand color from a palette
ggplot(mtcars, aes(hp, mpg)) +
geom_point(size = 4, color = msk_palette("main")[1])

# use a discrete color scale
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
geom_point(size = 4) +
scale_color_manual(values = msk_palette("contrast"))

# use a continuous color scale
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
geom_point(size = 4, alpha = .6) +
scale_color_gradientn(colors = msk_palette("blues", type = "continuous"))
}
\keyword{colors}
