% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_color_palette.R
\docType{data}
\name{msk_palettes}
\alias{msk_palettes}
\title{Complete list of MSK branded color palettes}
\format{An object of class \code{list} of length 4.}
\usage{
msk_palettes
}
\description{
Creates color palettes based on MSK brand colors, including
"main" for the main blue and orange colors, "blues" for all the blue colors,
"oranges" for all the orange colors, and "contrast" for an ordered palette of
contrasting colors.
}
\examples{
msk_palettes[["main"]]
msk_palettes[["contrast"]]
}
\keyword{datasets}
