% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvsurv.R
\name{uvsurv}
\alias{uvsurv}
\title{Table of univariable survival analysis results}
\usage{
uvsurv(contvars, catvars, event, time, test, dat)
}
\arguments{
\item{contvars}{is a list of the continuous variables you want in the rows
e.g. list('Age')}

\item{catvars}{is a list of the categorical variables you want in the rows
e.g. list('Gender','Race')}

\item{event}{is the survival event indicator (needs to be in quotes)}

\item{time}{is the survival time variable (needs to be in quotes)}

\item{test}{is the timepoint you would like to estimate, in whatever
units the survival time is in}

\item{dat}{is the dataset to use for analysis}
}
\value{
Returns a dataframe
}
\description{
\code{uvsurv} takes lists of continuous and/or categorical variables.
For continuous variables, \code{coxph} returns HR (95\% CI) and
log-rank p-values. For categorical variables, \code{coxph} returns
HR (95\% CI) and log-rank p-values and \code{survfit} produces median
survival (95\% CI) and a survival estimate at a specified time. Results
are put into  a table suitable for printing in a Word \code{R Markdown}
file.
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
